+++ mc-server-wrapper/src/main.rs
@@ -25,8 +25,8 @@ use crossterm::{
     terminal::{disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen},
     ExecutableCommand,
 };
+use ratatui::{backend::CrosstermBackend, Terminal};
 use structopt::StructOpt;
-use tui::{backend::CrosstermBackend, Terminal};
 use util::{format_online_players, OnlinePlayerFormat};
 
 mod config;