+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Notable `mc-server-wrapper` changes, tracked in the [keep a changelog](https://k
 * Updated `tui` to 0.19
 * Updated `crossterm` to 0.26.0
 * Updated `toml` to 0.7
-* Replaced `notify` with `notify-debouncer-mini` 0.2.1
+* Replaced `notify` with `notify-debouncer-mini` 0.3.0
 * Updated `textwrap` to 0.16.0
 * Updated other dependencies